requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
python-dotenv
gunicorn
psycopg2
